Improved inspect and added pretty printer to Coordinate class.
The existing inspect method now produces output roughly matching a traditional
inspect, but restricted to the x, y, and z instance variables, and truncated to
not output the full length of the floats.

A new pretty_print method has been added for use in things like the log system's
pretty printer. It outputs a much more terse version of the coordinate.