Reworked SMSG_UPDATE_OBJECT to not overwrite objects created before spawn.
Objects created through observation from packets are no longer overwritten by create object entries in
the SMSG_UPDATE_OBJECT payload. Instead, once an object appears in a create object entry, a call to spawn!
on the object is made, and appropriate state and a log item is added to the object.