Reworked DB2 parser to be less constrained to known structures.
In cases of files without documented structures, such as Creature.db2,
the DB2 parser was difficult to use as it relied on reading all fields
to advance through the file.

The parser now reads in a full record (as DB2 files define record
lengths in their headers), and uses the record count from the header
to instantiate all records, regardless of whether the structure is
known.

Additionally, a few issues related to index reading have been resolved.