Updated DBC version checking logic and added more version data
Since files can be identical across multiple versions, it doesn't
make sense to use file digests as keys unique to each file name in
the versions file.

We now store file digests as keys unique to each build, locale, and
filename (in that order).

This commit also lands the full set of DBC version hashes for the
following builds:

* 20338 enUS
* 20490 enUS