Relocated several definitions to build 19033
Supporting PTR and alpha / beta builds is not really in the cards
for wow-data, as it makes bounding various logic by build number
tricky.

As an example, build 18179 is an alpha build of Warlords of Draenor,
but its build number is lower than 18414, one of the last builds of
Mists of Pandaria.

As a result, we're going to treat all Warlords of Draenor era
definitions as starting in build 19033 (appears to be the first
retail build of WoD).